functions/sitemap-collector/requirements.txt
functions-framework==3.*
requests==2.31.0
xmltodict==0.13.0
feedparser==6.0.10
PyYAML==6.0.2
google-cloud==0.34.0
google-cloud-pubsub==2.28.0
sentry-sdk>=0.17.1
